change time stamp from str to double

+++ lib/Timed3dParameterBlock.h
@@ -68,7 +68,7 @@ class Timed3dParameterBlock : public SizedParameterBlock<3, 3, Eigen::Vector3d> 
   /// @param[in] id The (unique) ID of this block.
   /// @param[in] timestamp The timestamp of this state.
   Timed3dParameterBlock(const Eigen::Vector3d& point, uint64_t id, 
-                        const uint64_t timestamp) {
+                        const double timestamp) {
     setEstimate(point);
     setId(id);
     setTimestamp(timestamp);
@@ -89,7 +89,7 @@ class Timed3dParameterBlock : public SizedParameterBlock<3, 3, Eigen::Vector3d> 
   }
   
   /// @param[in] timestamp The timestamp of this state.
-  void setTimestamp(const uint64_t timestamp){timestamp_=timestamp;}
+  void setTimestamp(const double timestamp){timestamp_=timestamp;}
 
   // getters
   /// @brief Get estimate.
@@ -99,7 +99,7 @@ class Timed3dParameterBlock : public SizedParameterBlock<3, 3, Eigen::Vector3d> 
   }
   /// \brief Get the time.
   /// \return The timestamp of this state.
-  uint64_t timestamp() const {return timestamp_;}
+  double timestamp() const {return timestamp_;}
 
   // minimal internal parameterization
   // x0_plus_Delta=Delta_Chi[+]x0
@@ -142,7 +142,7 @@ class Timed3dParameterBlock : public SizedParameterBlock<3, 3, Eigen::Vector3d> 
   virtual std::string typeInfo() const {return "Timed3dParameterBlock";}
 
 private:
-  uint64_t timestamp_; ///< Time of this state.
+  double timestamp_; ///< Time of this state.
 };
 
 // } // namespace ceres
